+++ spec/forms/internal_contacts/edit_added_by_user_form_spec.rb
@@ -17,7 +17,7 @@ RSpec.describe InternalContacts::EditAddedByUserForm, type: :model do
 
   it "can be initialized with attributes" do
     user = create(:user, :regional_delivery_officer, email: "case.worker@education.gov.uk")
-    project = create(:conversion_project, regional_delivery_officer: user)
+    project = create(:conversion_project)
 
     result = described_class.new({email: user.email, project: project})
 
@@ -27,14 +27,6 @@ RSpec.describe InternalContacts::EditAddedByUserForm, type: :model do
 
   describe "validations" do
     describe "email" do
-      it "is required" do
-        project = create(:conversion_project, regional_delivery_officer: nil)
-
-        result = described_class.new_from_project(project)
-
-        expect(result).to be_invalid
-      end
-
       it "must be an @education.gov.uk address" do
         user = create(:user, :caseworker)
         project = create(:conversion_project, regional_delivery_officer: user)
@@ -69,7 +61,7 @@ RSpec.describe InternalContacts::EditAddedByUserForm, type: :model do
     context "when the form is valid" do
       it "updates the project and returns true" do
         user = create(:user, :caseworker)
-        project = create(:conversion_project, regional_delivery_officer: nil)
+        project = create(:conversion_project)
 
         result = described_class.new({email: user.email, project: project}).update
 
@@ -80,11 +72,12 @@ RSpec.describe InternalContacts::EditAddedByUserForm, type: :model do
 
     context "when the form is invalid" do
       it "does not update the project and returns false" do
-        project = create(:conversion_project, regional_delivery_officer: nil)
+        user = create(:user, :regional_delivery_officer)
+        project = create(:conversion_project, regional_delivery_officer: user)
 
         result = described_class.new({email: "invlalid@email.address", project: project}).update
 
-        expect(project.reload.regional_delivery_officer).to be_nil
+        expect(project.reload.regional_delivery_officer).to eql(user)
         expect(result).to be false
       end
     end