+++ spec/forms/transfer/create_project_form_spec.rb
@@ -353,10 +353,10 @@ RSpec.describe Transfer::CreateProjectForm, type: :model do
 
         context "when the group exists" do
           it "makes the association to the project" do
-            group = create(:project_group, group_identifier: "GRP_12345678", trust_ukprn: 1234567)
+            group = create(:project_group, group_identifier: "GRP_12345678", trust_ukprn: 12345678)
             form = build(
               :create_transfer_project_form,
-              incoming_trust_ukprn: 1234567,
+              incoming_trust_ukprn: 12345678,
               group_id: "GRP_12345678"
             )